+++ packages/taiko-client-rs/crates/test-harness/src/lib.rs
@@ -4,6 +4,7 @@ use std::sync::OnceLock;
 
 use tracing_subscriber::EnvFilter;
 
+mod beacon_stub;
 mod blob_server;
 mod helper;
 pub mod shasta;
@@ -11,6 +12,7 @@ pub mod shasta;
 #[cfg(feature = "preconfirmation")]
 pub mod preconfirmation;
 
+pub use beacon_stub::BeaconStubServer;
 pub use blob_server::BlobServer;
 pub use helper::{evm_mine, mine_l1_block};
 pub use shasta::{env::ShastaEnv, helpers::verify_anchor_block};
